Fix Typo & add link to full example
selecd -> select

Some feedback: IMO, the part that adds the axis should not be doing the selection of `#my-svg`, as in this example code `var axisG = d3.select('#my-svg').append('g');`. I was hoping this was an "on-the-way" example that you'd correct later on, but I don't see any mention of it. Perhaps after you introduce `selection.each`, you could give a modified version of the axis example that appends to the selection that's passed in, rather than selecting the svg.

Also, have you created a bl.ock with a complete example of the reusable charts pattern based on your example code here? The tutorial kind of leaves students hanging at the end, wondering how it can all be put together. I've added a link to a bl.ock I created with the code from @mbostock's original towards reusable charts example.